feat(drawzone): Gestion tactille
Changements apportés :
- Gestion des événements tactilles dans la drawzone
